backend/recommendation_service/app/requirements.txt
nameko
redispy
